pyproject.toml
[project]
name = "fcu-mcp"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "python-dotenv",
    "selenium",
    "fastmcp"
]
